Add sell nft commnad
Added the ability to sell your NFTs. Finn class for getting stock prices is now instantiated in every command because the library is only offered as a session that expires if the bot is not used for a couple minutes. AIOHTTP will replace finnhub library in the future to fix this issue and make async requests.